+++ packages/cem-inheritance/src/cem-inheritance.ts
@@ -1,5 +1,5 @@
 import { CEM, Component } from "../../../tools/cem-utils/index.js";
-import { createOutDir, logBlue, saveFile } from "../../../tools/integrations";
+import { createOutDir, logBlue, logYellow, saveFile } from "../../../tools/integrations";
 import { Options } from "./types.js";
 
 const completedClasses: string[] = [];
@@ -10,7 +10,11 @@ let externalComponents: Component[] = [];
 let updatedCEM: any = {};
 
 export function updateCemInheritance(cem: CEM, options: Options = {}) {
-  logBlue("[cem-inheritance-generator] - Updating Custom Elements Manifest...");
+  if(options.skip) {
+    logYellow("[cem-inheritance-generator] - Skipped", options.hideLogs)
+  }
+  
+  logBlue("[cem-inheritance-generator] - Updating Custom Elements Manifest...", options.hideLogs);
   const newCem = generateUpdatedCem(cem, options);
   createOutDir(userConfig.outdir!);
   saveFile(
@@ -18,7 +22,7 @@ export function updateCemInheritance(cem: CEM, options: Options = {}) {
     userConfig.fileName!,
     JSON.stringify(newCem, null, 2)
   );
-  logBlue("[cem-inheritance-generator] - Custom Elements Manifest updated.");
+  logBlue("[cem-inheritance-generator] - Custom Elements Manifest updated.", options.hideLogs);
 }
 
 function updateOptions(options: Options = {}) {