+++ packages/expanded-types/src/cem-plugin.ts
@@ -1,9 +1,15 @@
 import path from "path";
 import fs from "fs";
-import type { CEM, Component } from "../../../tools/cem-utils";
+import type { Component } from "../../../tools/cem-utils";
+import { logBlue, logYellow } from "../../../tools/integrations";
 
 export interface Options {
+  /** Determines the name of the property used in the manifest to store the expanded type */
   propertyName?: string;
+  /** Hides logs produced by the plugin */
+  hideLogs?: boolean;
+  /** Prevents plugin from executing */
+  skip?: boolean;
 }
 
 interface AliasTypes {
@@ -48,6 +54,14 @@ export function expandTypesPlugin(
   }
 ) {
   options = op;
+  if (options.skip) {
+    logYellow("[cem-expanded-types] - Skipped", options.hideLogs);
+    return;
+  }
+  logBlue(
+    "[cem-expanded-types] - Updating Custom Elements Manifest...",
+    options.hideLogs
+  );
 
   return {
     name: "expand-types-plugin",
@@ -74,7 +88,10 @@ export function getTsProgram(
     configName
   );
   const { config } = ts.readConfigFile(tsConfigFile, ts.sys.readFile);
-  const compilerOptions = ts.convertCompilerOptionsFromJson(config.compilerOptions ?? {}, ".");
+  const compilerOptions = ts.convertCompilerOptionsFromJson(
+    config.compilerOptions ?? {},
+    "."
+  );
   const program = ts.createProgram(globs, compilerOptions.options);
   typeChecker = program.getTypeChecker();
   return program;
@@ -179,8 +196,6 @@ function groupTypesByName() {
       groupedTypes[type][alias] = aliasTypes[alias][type];
     }
   }
-
-  // console.log("groupedTypes", JSON.stringify(groupedTypes, null, 2));
 }
 
 function setEnumTypes(node: any) {
@@ -239,6 +254,10 @@ function analyzePhase({ ts, node, moduleDoc, context }: any) {
   }
 
   updateExpandedTypes(component, context);
+  logBlue(
+    "[cem-expanded-types] - Custom Elements Manifest updated.",
+    options.hideLogs
+  );
 }
 
 function getComponent(node: any, moduleDoc: any) {