+++ packages/svelte-integration/src/type-generator.ts
@@ -6,10 +6,19 @@ import {
   getCustomEventTypes,
   getMemberDescription,
 } from "../../../tools/cem-utils";
-import { createOutDir, logBlue, saveFile } from "../../../tools/integrations";
+import { createOutDir, logBlue, logYellow, saveFile } from "../../../tools/integrations";
 import { Options } from "./types";
 
 export function generateSvelteTypes(manifest: any, options: Options) {
+  if (options.skip) {
+    logYellow("[svelte-type-generator] - Skipped", options.hideLogs);
+    return;
+  }
+  logBlue(
+    "[svelte-type-generator] - Updating Custom Elements Manifest...",
+    options.hideLogs
+  );
+  
   options = getOptions(options);
 
   const components = getComponents(manifest, options.exclude).filter(
@@ -24,7 +33,7 @@ export function generateSvelteTypes(manifest: any, options: Options) {
     "typescript",
     120
   );
-  logBlue(`[svelte-type-generator] - Generated "${outputPath}".`);
+  logBlue(`[svelte-type-generator] - Generated "${outputPath}".`, options.hideLogs);
 }
 
 function getOptions(options: Options) {