+++ tools/configurations/src/config.ts
@@ -18,9 +18,13 @@ export interface BaseOptions {
   /** Displays the methods section of the element description */
   hideMethodDocs?: boolean;
   /** Overrides the default section labels in the component description */
-  labels?: DescriptionLabels;  
+  labels?: DescriptionLabels;
   /** The property form your CEM component object to display your types */
   typesSrc?: string;
+  /** Hides logs produced by the plugin */
+  hideLogs?: boolean;
+  /** Prevents plugin from executing */
+  skip?: boolean;
 }
 
 export interface DescriptionLabels {