+++ packages/jet-brains-integration/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.5.0
+
+- Added the ability to hide logs
+- Added the ability to skip the plugin
+
 ## 1.4.4
 
 - Allow any string for `descriptionSrc`