+++ packages/solidjs-integration/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.7.0
+
+- Added the ability to hide logs
+- Added the ability to skip the plugin
+
 ## 1.6.1
 
 - Allow any string for `descriptionSrc`